22-5-5 arreglado borrar imagen 1 en update

+++ laravel/Projecte/app/Http/Controllers/BlogController.php
@@ -160,10 +160,10 @@ class BlogController extends Controller
 
             if (\Storage::disk('public')->exists($filePath)) {
                 $fullPath = \Storage::disk('public')->path($filePath);
-
-                \Storage::disk('public')->delete($imatge->filepath);
-                //$imatge->delete();
-
+                if($imatge->id != 1) {
+                    \Storage::disk('public')->delete($imatge->filepath);
+                    //$imatge->delete();
+                }
                 $imagen = Image::create([
                     'filepath' => $filePath,
                     'filesize' => $fileSize
